requirements.txt
# Core Application
streamlit==1.45.1
google-generativeai==0.8.5
python-dotenv==1.1.0

# Testing
pytest==8.3.5
pytest-mock==3.14.0

# Cloud Run (production server)
gunicorn==23.0.0
